requirements.txt
flask
joblib
numpy
requests
Pillow
torch
torchvision
scikit-learn
